Fixed a problem with the gui objects for PD / Linux.
o.display was crashing when o.compose was doing a refresh.
It turns out that o.display was getting called with o.compose's object,
and it seems to be due to the fact that both object defined a function
called jbox_redraw that was just a wrapper around their own redrawing
code. Geting jbox_redraw out of there fixed everything